Convert template to block style and rename theme to scheme
Hugo appears to have grown the use of blocks rather than partials.
Therefore, this is a conversion to using {{ block ...}} type templates
and reducing the partial to the sidebar.  A future change may remove
that too.

Also, due to a really obscure bug, the color theme in the theme is being
renamed from 'theme' to 'scheme' in the params section.  See the
README.md for details.